fix: restore libxposed api 100 compatibility
