+++ app/src/main/kotlin/eu/hxreborn/remembermysort/RememberMySortModule.kt
@@ -1,33 +1,36 @@
 package eu.hxreborn.remembermysort
 
 import android.database.Cursor
-import android.util.Log
 import eu.hxreborn.remembermysort.hook.DirectoryLoaderHooker
 import eu.hxreborn.remembermysort.hook.FolderLoaderHooker
 import eu.hxreborn.remembermysort.hook.LongPressHooker
 import eu.hxreborn.remembermysort.hook.RecentsLoaderHooker
 import eu.hxreborn.remembermysort.hook.SortCursorHooker
+import eu.hxreborn.remembermysort.hook.SortDialogDismissHooker
 import io.github.libxposed.api.XposedInterface
 import io.github.libxposed.api.XposedModule
 import io.github.libxposed.api.XposedModuleInterface.ModuleLoadedParam
-import io.github.libxposed.api.XposedModuleInterface.PackageReadyParam
+import io.github.libxposed.api.XposedModuleInterface.PackageLoadedParam
 
 internal lateinit var module: RememberMySortModule
 
-class RememberMySortModule : XposedModule() {
-    override fun onModuleLoaded(param: ModuleLoadedParam) {
+class RememberMySortModule(
+    base: XposedInterface,
+    param: ModuleLoadedParam,
+) : XposedModule(base, param) {
+    init {
         module = this
-        log(Log.INFO, TAG, "v${BuildConfig.VERSION_NAME} loaded in ${param.processName}")
+        log("v${BuildConfig.VERSION_NAME} loaded")
     }
 
-    override fun onPackageReady(param: PackageReadyParam) {
+    override fun onPackageLoaded(param: PackageLoadedParam) {
         if (!param.isFirstPackage) return
 
         hookSortCursor(param.classLoader)
         hookSortListFragment(param.classLoader)
         hookLoaders(param.classLoader)
 
-        log(Log.INFO, TAG, "Module initialized in ${param.packageName}")
+        log("Module initialized in ${param.packageName}")
     }
 
     private fun hookSortCursor(classLoader: ClassLoader) {
@@ -35,8 +38,10 @@ class RememberMySortModule : XposedModule() {
         runCatching {
             val sortModel = classLoader.loadClass(className)
             val lookup = classLoader.loadClass("com.android.documentsui.base.Lookup")
-            hook(sortModel.getDeclaredMethod("sortCursor", Cursor::class.java, lookup))
-                .intercept(SortCursorHooker)
+            hook(
+                sortModel.getDeclaredMethod("sortCursor", Cursor::class.java, lookup),
+                SortCursorHooker::class.java,
+            )
             log("Hooked $className.sortCursor")
         }.onFailure { e ->
             log("Failed to hook $className.sortCursor", e)
@@ -47,16 +52,8 @@ class RememberMySortModule : XposedModule() {
         for (className in SORT_FRAGMENT_CLASSES) {
             runCatching {
                 val clazz = classLoader.loadClass(className)
-                hook(clazz.getMethod("onStart")).intercept { chain ->
-                    val result = chain.proceed()
-                    LongPressHooker.handleOnStart(chain.thisObject)
-                    result
-                }
-                hook(clazz.getMethod("onStop")).intercept { chain ->
-                    val result = chain.proceed()
-                    LongPressHooker.clearDialogState()
-                    result
-                }
+                hook(clazz.getMethod("onStart"), LongPressHooker::class.java)
+                hook(clazz.getMethod("onStop"), SortDialogDismissHooker::class.java)
                 log("Hooked $className.onStart/onStop")
             }.onFailure {
                 log("$className not found, skipping")
@@ -68,7 +65,7 @@ class RememberMySortModule : XposedModule() {
         for ((className, hooker) in LOADERS) {
             runCatching {
                 val loaderClass = classLoader.loadClass(className)
-                hook(loaderClass.getDeclaredMethod("loadInBackground")).intercept(hooker)
+                hook(loaderClass.getDeclaredMethod("loadInBackground"), hooker)
                 log("Hooked $className.loadInBackground")
             }.onFailure {
                 log("$className not found, skipping")
@@ -77,26 +74,22 @@ class RememberMySortModule : XposedModule() {
     }
 
     companion object {
-        const val TAG = "RememberMySort"
+        private val SORT_FRAGMENT_CLASSES = listOf(
+            "com.android.documentsui.sorting.SortListFragment",
+            "com.google.android.documentsui.sorting.SortListFragment",
+        )
 
-        private val SORT_FRAGMENT_CLASSES =
-            listOf(
-                "com.android.documentsui.sorting.SortListFragment",
-                "com.google.android.documentsui.sorting.SortListFragment",
-            )
-
-        private val LOADERS: List<Pair<String, XposedInterface.Hooker>> =
-            listOf(
-                "com.android.documentsui.DirectoryLoader" to DirectoryLoaderHooker,
-                "com.android.documentsui.loaders.FolderLoader" to FolderLoaderHooker,
-                "com.android.documentsui.RecentsLoader" to RecentsLoaderHooker,
-            )
+        private val LOADERS: List<Pair<String, Class<out XposedInterface.Hooker>>> = listOf(
+            "com.android.documentsui.DirectoryLoader" to DirectoryLoaderHooker::class.java,
+            "com.android.documentsui.loaders.FolderLoader" to FolderLoaderHooker::class.java,
+            "com.android.documentsui.RecentsLoader" to RecentsLoaderHooker::class.java,
+        )
 
         fun log(
             msg: String,
             t: Throwable? = null,
         ) {
-            if (t != null) module.log(Log.ERROR, TAG, msg, t) else module.log(Log.INFO, TAG, msg)
+            if (t != null) module.log(msg, t) else module.log(msg)
         }
     }
 }
